% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABLXtag.R
\name{animap}
\alias{animap}
\title{Fucntion to animate the results of the itersf function}
\usage{
animap(anidat, iter)
}
\arguments{
\item{anidat}{iterout$modeliterations (pulls the list item modeliterations from the itersf output, here called "iterout")}

\item{iter}{Number of iterations}
}
\description{
Fucntion to animate the results of the itersf function
}

