% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABLXtag.R
\name{CI2shp_boot}
\alias{CI2shp_boot}
\title{Creates confidence interval polygons based on results of bootstrap iterations}
\usage{
CI2shp_boot(track, fname = "testshp", level = 0.95, npoints = 100,
  proj4string = CRS("+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs "))
}
\arguments{
\item{track}{data.frame of $iter_summary from the result of itersf}

\item{fname}{output file name, stored in workspace, not in folder}

\item{level}{Confidence level}

\item{npoints}{Number of points to include in polygon}

\item{proj4string}{Projection for polygon}
}
\description{
Creates confidence interval polygons based on results of bootstrap iterations
}

