% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABLXtag.R
\name{get.bath.data}
\alias{get.bath.data}
\title{Dowload bathymetric coverage}
\usage{
get.bath.data(lonlow, lonhigh, latlow, lathigh, folder = tempdir(),
  seaonly = T, res = c(0.5, 1))
}
\arguments{
\item{lonhigh}{bounding longitude}

\item{latlow}{bounding latitude}

\item{lathigh}{bounding latitude}

\item{folder}{directory, defaults to tempdir()}

\item{seaonly}{gets rid of land topography}

\item{res}{resolution of output data, here either 1 or 0.5, larger number is smaller file size, low resolution}

\item{longlow}{bounding longitude}
}
\description{
Dowload bathymetric coverage
}

