% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABLXtag.R
\name{getMWTarch}
\alias{getMWTarch}
\title{Function to extract MWT archived tag data from excel workbooks, formats data for 
running models}
\usage{
getMWTarch(xlsfile, tagID, tyear, taglocfile, dayT = NULL, xT = NULL)
}
\arguments{
\item{xlsfile}{Path to the excell workboot for tagID}

\item{tagID}{ARGOS ID of the tag being run}

\item{tyear}{Year which the tag was deployed, for use in looping through tags}

\item{dayT}{Date of last transmission from .getLastDay}

\item{xT}{Location of last transmission from .getLastLoc}

\item{tagLocfile}{Path to the .txt file listing all tags release information}
}
\description{
Function to extract MWT archived tag data from excel workbooks, formats data for 
running models
}

