% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABLXtag.R
\name{itersf}
\alias{itersf}
\title{Function that repeats kftrack and btrack through multiple iterattions and summarizes output}
\usage{
itersf(iter, xtrack, kf.dir)
}
\arguments{
\item{iter}{Number of iterations}

\item{xtrack}{Output from prepf_arch function}

\item{kf.dir}{Path to location of admb folder}
}
\description{
Function that repeats kftrack and btrack through multiple iterattions and summarizes output
}

