% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABLXtag.R
\name{kftrack}
\alias{kftrack}
\title{Kalman filter model estimation of geolocation, wrapper for admb function}
\usage{
kftrack(kfdata, fix.first = TRUE, fix.last = TRUE,
  theta.active = c(u.active, v.active, D.active, bx.active, by.active,
  sx.active, sy.active, a0.active, b0.active, vscale.active),
  theta.init = c(u.init, v.init, D.init, bx.init, by.init, sx.init, sy.init,
  a0.init, b0.init, vscale.init), u.active = TRUE, v.active = TRUE,
  D.active = TRUE, bx.active = TRUE, by.active = TRUE, sx.active = TRUE,
  sy.active = TRUE, a0.active = TRUE, b0.active = TRUE,
  vscale.active = TRUE, u.init = 0, v.init = 0, D.init = 100,
  bx.init = 0, by.init = 0, sx.init = 0.5, sy.init = 1.5,
  a0.init = 0.001, b0.init = 0, vscale.init = 1,
  var.struct = "solstice", dev.pen = 0, save.dir = NULL,
  admb.string = "", kf.dir)
}
\arguments{
\item{kfdata}{from the output of prepf_arch}

\item{fix.first}{logical if the model should leave the first (i.e., tagging location) location fixed}

\item{fix.last}{logical if the model should leave the last (i.e., recovery or last transmission) location fixed}

\item{theta.active}{admb inputs}

\item{theta.init}{admb inputs}

\item{kf.dir}{Path to location of admb folder}

\item{file}{admb .dat file}
}
\description{
Kalman filter model estimation of geolocation, wrapper for admb function
}

