% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABLXtag.R
\name{make.btrack}
\alias{make.btrack}
\title{Background function that operates within make.btrack}
\usage{
make.btrack(fmat, bathy, save.samp = F, mintype = 2, ci = 0.95,
  npoints = 300, fulldist = T)
}
\arguments{
\item{fmat}{results of prepb}

\item{bathy}{loaded bathymetric coverage}

\item{save.samp}{Should the sampled points be saved, makes for much larger file}

\item{mintype}{not exactly sure about this, it's on my list to investigate further}

\item{ci}{confidence interval}

\item{npoints}{number of points to resample}

\item{fulldist}{should the full distribution of points be used in variance calcs}
}
\description{
Background function that operates within make.btrack
}

