% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABLXtag.R
\docType{data}
\name{month.colors}
\alias{month.colors}
\title{Object with monthly colors, can be adapted as desired}
\format{An object of class \code{matrix} with 12 rows and 2 columns.}
\usage{
month.colors
}
\description{
Object with monthly colors, can be adapted as desired
}
\keyword{datasets}

