% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABLXtag.R
\name{plot.btrack}
\alias{plot.btrack}
\title{Plots the bathymetric corrected estimated locations}
\usage{
plot.btrack(btrack, cex = 1.5, ci = F, add = F, alpha = 0.15,
  bymonth = T, pch = 21, bg = 4, legend = F)
}
\arguments{
\item{btrack}{results of make.btrack}

\item{cex}{adjusts size of points}

\item{add}{logical to add points}

\item{alpha}{graphical parameter}

\item{bymonth}{logical if the points should be colored by month}

\item{pch}{point type}

\item{bg}{device background color}

\item{legend}{logical if legend should be included}
}
\description{
Plots the bathymetric corrected estimated locations
}

