% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABLXtag.R
\name{prepb}
\alias{prepb}
\title{Prepares kftrack results for the btrack function}
\usage{
prepb(kfit, prepf, span = NULL)
}
\arguments{
\item{kfit}{kftrack results}

\item{prepf}{prepf_arch results}

\item{span}{passed to internal function}
}
\description{
Prepares kftrack results for the btrack function
}

