% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABLXtag.R
\name{prepf_arch}
\alias{prepf_arch}
\title{Function cleans up tag data to prepare it to run in kftrack}
\usage{
prepf_arch(tag, xmin = -100, xmax = 0, ymin = 10, ymax = 55,
  keepall = F, sst.depth = NULL, use.minmax = F)
}
\arguments{
\item{tag}{output file from getMWTarch}

\item{xmin}{minimum bounding longitude}

\item{xmax}{maximum bounding longitude}

\item{ymin}{minimum bounding latitude}

\item{ymax}{maximum bounding latitude}

\item{keepall}{logical to keep values outside of bounds}

\item{sst.depth}{not sure what this does, we aren't using SST}

\item{use.minmax}{logical to use x/y min/max}
}
\description{
Function cleans up tag data to prepare it to run in kftrack
}

